+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- Replace `account_id` with `institution_id` in `ImportSessions` table
 - Allow editing Asset Class in Asset SubClass popup
 - Restyled Institutions maintenance window for consistent look and feel
 - Fix compile error in Institutions view due to missing empty state component